requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
stripe==7.10.0
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
